fix(comment-crud):Only comment owner can delete their comment only

+++ src/main/java/com/artichourey/insighthub/serviceImpl/CommentServiceImpl.java
@@ -12,6 +12,7 @@ import com.artichourey.insighthub.exception.ResourceNotFoundException;
 import com.artichourey.insighthub.mapper.CommentMapper;
 import com.artichourey.insighthub.repositories.CommentRepository;
 import com.artichourey.insighthub.repositories.PostRepository;
+import com.artichourey.insighthub.repositories.UserRepository;
 import com.artichourey.insighthub.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,12 +26,16 @@ public class CommentServiceImpl implements CommentService {
 	private final CommentRepository commentRepository;
     private final PostRepository postRepository;
     private final CommentMapper commentMapper;
+    private final UserRepository userRepository;
 
     @Override
-    public CommentResponseDto addComment(Long postId, CommentRequestDto request, User user) {
+    public CommentResponseDto addComment(Long postId, CommentRequestDto request, String username) {
 
-        log.info("Adding comment: postId={}, userId={}, parentCommentId={}, content={}",
-                postId, user.getId(), request.getParentCommentId(), request.getContent());
+        log.info("Adding comment: postId={}, username={}, parentCommentId={}, content={}",
+                postId, username, request.getParentCommentId(), request.getContent());
+
+        User user = userRepository.findByName(username)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ResourceNotFoundException("Post not found"));
@@ -43,10 +48,6 @@ public class CommentServiceImpl implements CommentService {
         }
 
         Comment comment = commentMapper.toEntity(request, user, post, parent);
-
-        comment.setPost(post);
-        comment.setUser(user);
-        comment.setParentComment(parent);
 
         Comment savedComment = commentRepository.save(comment);
 
@@ -64,11 +65,15 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId,String name) {
     	log.info("Deleting comment: commentId={}", commentId);
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
+        
+        if(!comment.getUser().getName().equals(name)){
+            throw new RuntimeException("You are not allowed to delete this comment");
+        }
         commentRepository.delete(comment); 
         log.info("Comment deleted successfully: commentId={}", commentId);
     }
